Updated paths and second password for LUKS encryption

+++ disko/disko-config.sh
@@ -6,14 +6,22 @@ read DEVICENAME
 echo "Enter your desired LUKS password:"
 read -s LUKSPASS
 
+echo
+read -s -p "Confirm LUKS password: " LUKSPASS2
+echo
+if [ "$LUKSPASS" != "$LUKSPASS2" ]; then
+  echo "Passwords do not match!" >&2
+  exit 1
+fi
+
 # Write password to a temporary keyfile
 echo -n "$LUKSPASS" > /tmp/secret.key
 
 # Copy and modify template disko-config, replacing device name
-sed "s|/dev/mydisk|$DEVICENAME|g" ./disko-config-template.nix > ./disko-config.nix
+sed "s|/dev/mydisk|$DEVICENAME|g" ./disko/disko-config-template.nix > ./disko-config.nix
 
 # Trigger disko with supplied config and password
-sudo nix run github:nix-community/disko -- --mode disko ./disko-config.nix
+sudo nix --extra-experimental-features 'nix-command flakes' run github:nix-community/disko -- --mode disko ./disko-config.nix
 
 # Remove temporary keyfile after partitioning for security
 rm -f /tmp/secret.key
